--- a/ModBuilder/generalsmodbuilder/__version__.py
+++ b/ModBuilder/generalsmodbuilder/__version__.py
@@ -1,3 +1,3 @@
-VERSION = (1, 8)
+VERSION = (1, 9)
 
 __version__ = '.'.join(map(str, VERSION))