These changes increase build performance significantly.

GeneralsGamePatch as of 09-03-2023:

A no-change rebuild for Full English pack is reduced from 20 to 4 seconds (in debugger).

A complete build for Full English pack is reduced from 187 to 82 seconds (in debugger).


Additionally a change log generation feature was added.